fjernet (int)

+++ src/com/company/Main.java
@@ -152,9 +152,9 @@ public class Main {
 
                 if (player1.getBalance() >= 40){
                     System.out.println("over 40 point");
-                    System.out.println((int)die1+ " " + (int)die2);
+                    System.out.println(die1+ " " + (int)die2);
 
-                    if ((int)die1 == (int)die2){
+                    if (die1 == die2){
                         System.out.println("spiller 1 har vundet!");
                         gui.showMessage(player1.getName() + " du har vundet!");
                         hasPlayerOneWon = true;
